Refactoring to the latest versions of numpy and matplotlib

+++ PyP-BEAGLE/beagle_mock_catalogue.py
@@ -159,7 +159,7 @@ class BeagleMockCatalogue(object):
 
     def get_param_values(self, ID, names):
 
-        values = np.zeros(len(names), dtype=np.float32)
+        values = np.zeros(len(names), dtype=float)
 
         for i, name in enumerate(names):
